break out test_path_creation
