+++ trimesh/version.py
@@ -1,1 +1,1 @@
-__version__ = '3.6.30'
+__version__ = '3.6.31'
